# iris-api/openapi.yaml
openapi: 3.1.0
info:
  title: Iris Operations API
  version: 0.3.0
  description: |
    Shared HTTP API for Iris work-order operations. Production runtime persists
    data in SQLite, while fixture-backed data remains available for development
    and automated tests.
servers:
  - url: http://localhost:8080
...
paths:
...
  /catalog-items/{id}:
    get:
      tags: [Catalog]
      summary: Return a single catalog item by id
      operationId: getCatalogItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The catalog item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItem'
        '404':
          description: Catalog item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Catalog]
      summary: Update a catalog item
      description: >
        Admins update every field. Operators (non-admin) may change only the
        item's kind (vrsta); all other fields in the payload are ignored and the
        stored values are preserved. Operators cannot create items.
      operationId: updateCatalogItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogItemInput'
      responses:
        '200':
          description: Updated catalog item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItem'
        '403':
          description: Operator attempted to create a catalog item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Catalog]
      summary: Delete a catalog item (admin only)
      operationId: deleteCatalogItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deletion result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '403':
          description: Caller is not an administrator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    IrisSession:
      type: apiKey
      in: cookie
      name: iris_session
...
  schemas:
...
    CatalogItem:
      type: object
      required: [id, code, name, kind, unit, salePrice, barcode, taxGroup, description, isActive]
      properties:
        id:
          type: string
        code:
          type: string
          description: Stable item code (legacy SIFRA on imported rows).
        name:
          type: string
        kind:
          type: string
          enum: [service, article]
        unit:
          type: string
          description: Unit of measure, lowercased (e.g. kom, set, m2, tab, ris).
          example: kom
        purchasePrice:
          type: [number, 'null']
          format: double
          description: >-
            Cost figure: purchase price (nabavna cena) for articles, cost of
            labor (cena rada) for services. Admin-only — the API returns null
            for non-admin users so operators never see cost or margin data.
        salePrice:
          type: [number, 'null']
          format: double
          description: Sale price charged to the customer (prodajna cena).
        barcode:
          type: [string, 'null']
        taxGroup:
          type: [string, 'null']
        description:
          type: [string, 'null']
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CatalogItemInput:
      type: object
      required: [name, kind, unit]
      properties:
        code:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [service, article]
        unit:
          type: string
          example: kom
        purchasePrice:
          type: [number, 'null']
          format: double
          description: Cost (nabavna cena / cena rada). Admin-only.
        salePrice:
          type: [number, 'null']
          format: double
          description: Sale price (prodajna cena).
        barcode:
          type: [string, 'null']
        taxGroup:
          type: [string, 'null']
        description:
          type: [string, 'null']
        isActive:
          type: boolean
        effectiveFrom:
          type: [string, 'null']
          format: date
          description: >-
            Date (YYYY-MM-DD) a changed price takes effect. Optional; defaults to
            today. Must be today or later — past dates are rejected. A future date
            schedules the new price without changing the item's displayed current
            price until that date arrives.
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: >-
            Serbian message safe to show the operator. Internal detail (SQL,
            file paths, decode errors) is logged and reported to Sentry
            instead of being returned here.
        requestId:
          type: string
          description: >-
            Short reference for this failed request, also sent as the
            X-Request-Id header and attached to the server log line and the
            Sentry event. Clients show it with the error message so a
            photographed screen can be traced to the exact request.